+++ frontend/src/app/modules/shared/assembler/notification.assembler.spec.ts
@@ -71,6 +71,7 @@ describe('InvestigationsAssembler', () => {
             sendTo: '',
             reason: { close: '', accept: '', decline: '' },
             assetIds: [],
+            errorMessage: ''
           },
           {
             id: 'test-2',