+++ CHANGELOG.md
@@ -9,12 +9,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - OAuth2 client credentials rest template interceptor
 - Configuration for left and right policies to use registry client library
+- Add support for JustInSequence aspect model
+- TRG Github Action Pipeline for quality checks
+- ErrorMessage field to investigation and alerts
 
 ### Changed
 
 - Changed digitalTwinRegistryRestTemplate to use token in requests
 - Update asBuild test data to 1.5.3 and asPlanned to 1.5.1
 - Changed transfer notification logic not to break iteration loop when sending notifications to bpn with more than 1 connector endpoints
+- Decentral flow only using bpn for resolving globalAssetIds
+- Logic of commitId retrieval by frontend has changed
+- Readme / Docker Notice information as required in TRGs
+- Removed references to GitHub registry and added docker hub for tractusx instead
 
 ### Removed